moved all authorization extension logic from space to baseChallenge
